yolov5-coreml-tflite-converter/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
-e git+ssh://git@github.com/SchweizerischeBundesbahnen/sbb-ml-models.git@485356aeb0a277907c160d435f7f654154046a70#egg=converter_utils&subdirectory=yolov5-coreml-tflite-converter/converter_utils
coremltools==5.2.0
cycler==0.11.0
flatbuffers==1.12
fonttools==4.32.0
gast==0.3.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
mpmath==1.2.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pluggy==1.0.0
protobuf==3.18.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
six==1.15.0
sounddevice==0.4.4
sympy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
-e git+ssh://git@github.com/SchweizerischeBundesbahnen/sbb-ml-models.git@485356aeb0a277907c160d435f7f654154046a70#egg=tflite&subdirectory=yolov5-coreml-tflite-converter/tflite
tflite-support-nightly==0.3.0.dev20220420
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.12.1
-e git+https://github.com/ultralytics/yolov5@956be8e642b5c10af4a1533e09084ca32ff4f21f#egg=yolov5
zipp==3.8.0
